fix: support high signal filter for lens notifications
