+++ src/components/Notification/NotificationSettings.tsx
@@ -12,42 +12,10 @@ import SettingsIcon from '@/assets/setting.svg';
 import { type SocialSource, Source } from '@/constants/enum.js';
 import { FireflySocialMediaProvider } from '@/providers/firefly/SocialMedia.js';
 import { NotificationPlatform, NotificationPushType, NotificationTitle } from '@/providers/types/Firefly.js';
+import { useNotificationSettings } from '@/hooks/useNotificationSettings.js';
 
 export function NotificationSettings({ source }: { source: SocialSource }) {
-    const { data, isLoading, refetch } = useQuery({
-        queryKey: ['notification-push-switch'],
-        async queryFn() {
-            return FireflySocialMediaProvider.getNotificationPushSwitch();
-        },
-    });
-
-    const enabled = useMemo(() => {
-        const item = data?.list.find((x) => x.title === NotificationTitle.NotificationsMode);
-
-        switch (source) {
-            case Source.Farcaster:
-                return (
-                    item?.list.find(
-                        (x) =>
-                            x.platform === NotificationPlatform.Priority &&
-                            x.push_type === NotificationPushType.Priority,
-                    )?.state ?? false
-                );
-            case Source.Lens:
-                return (
-                    item?.list.find(
-                        (x) =>
-                            x.platform === NotificationPlatform.Priority && x.push_type === NotificationPushType.Lens,
-                    )?.state ?? false
-                );
-            case Source.Twitter:
-                return false;
-            default:
-                safeUnreachable(source);
-                return false;
-        }
-    }, [data?.list, source]);
-
+    const { enabled, isLoading, refetch } = useNotificationSettings(source);
     const [{ loading }, onSwitch] = useAsyncFn(
         async (state: boolean) => {
             const pushTypes: Record<SocialSource, NotificationPushType | undefined> = {
